+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Redis function `FLUSHALL` with implementation defined synchrony ([#16])
+- Redis function `FLUSHALL` with implementation defined synchrony ([#17])
 
 ## [0.5.0] - 2024-03-06
 
@@ -65,4 +65,4 @@
 [#3]: https://github.com/JosuaKrause/redipy/pull/3
 [#7]: https://github.com/JosuaKrause/redipy/pull/7
 [#15]: https://github.com/JosuaKrause/redipy/pull/15
-[#16]: https://github.com/JosuaKrause/redipy/pull/16
+[#17]: https://github.com/JosuaKrause/redipy/pull/17